contrib/data-share/provider_example/requirements.txt
Django==2.0.2
djangorestframework==3.7.7
python-magic==0.4.15
transmissionrpc==0.11
